use float instead of double

+++ Transforms/binningtransform.cpp
@@ -48,14 +48,14 @@ private:
     {
         for ( uint32_t ch = 0; ch < channelCount; ++ch )
         {
-            double sum = 0.0;
+            float sum = 0.0;
             for ( int i = 0; i < _bin.height; ++i )
                 for ( int j = 0; j < _bin.width; ++j )
                 {
                     sum += pSrcPixel[( this->_pSrcBitmap->GetWidth() * i + j ) * channelCount + ch];
                 }
             sum /= _bin.width * _bin.height;
-            pDstPixel[ch] = static_cast< ChannelType >( std::clamp<double>( sum, 0, std::numeric_limits<ChannelType>::max() ) );
+            pDstPixel[ch] = static_cast< ChannelType >( std::clamp<float>( sum, 0, std::numeric_limits<ChannelType>::max() ) );
         }
     }
